HSI: omap_ssi: Introducing OMAP SSI driver
Introduces the OMAP SSI driver in the kernel.

The Synchronous Serial Interface (SSI) is a legacy version
of HSI. As in the case of HSI, it is mainly used to connect
Application engines (APE) with cellular modem engines (CMT)
in cellular handsets.

It provides a multichannel, full-duplex, multi-core communication
with no reference clock. The OMAP SSI block is capable of reaching
speeds of 110 Mbit/s.

Change-Id: Iea002a8f321cf66ab439552620af001206626ed6
Signed-off-by: Carlos Chinea <carlos.chinea@nokia.com>
Reviewed-on: http://gerrit.lud.stericsson.com/gerrit/20576
Reviewed-by: Pawel SZYSZUK <pawel.szyszuk@stericsson.com>
Tested-by: Pawel SZYSZUK <pawel.szyszuk@stericsson.com>
Reviewed-by: Jonas ABERG <jonas.aberg@stericsson.com>