ARM: twd: Adjust localtimer frequency with cpufreq notifiers
The clock to the ARM TWD local timer scales with the cpu
frequency. To allow the cpu frequency to change while
maintaining a constant TWD frequency, pick a lower target
frequency for the TWD and use the prescaler to divide down
to the closest lower frequency.

This patch provides a new initialization function that takes
a target TWD frequency and the ratio between the cpu
clock and the TWD clock, specified as an integer divider >= 2
in the Cortex A9 MPCore TRM, and 2 in the ARM11 MPCore TRM.
It also registers a cpufreq notifier that adjusts the
prescaler when the cpu frequency changes.

ST-Ericsson Linux next: Colin is driving this
ST-Ericsson ID: 279802
ST-Ericsson FOSS-OUT ID: Trivial

Change-Id: Ieec17963da8d7d37d935782633406bc1a0484dc9
Signed-off-by: Colin Cross <ccross@android.com>
Reviewed-on: http://gerrit.lud.stericsson.com/gerrit/11532
Tested-by: Martin PERSSON <martin.persson@stericsson.com>
Reviewed-by: QATOOLS
Reviewed-by: Jonas ABERG <jonas.aberg@stericsson.com>
Signed-off-by: Lee Jones <lee.jones@linaro.org>