mmc: Fixup wakelock handling
Wakelocks was not used correctly in mmc core.
A suspend operation could be triggered even though
there was an ongoing mmc operation.

Instead now one wakelock exist per mmc host. Once the host
is enabled the wakelock is retrieved. When the host is ready
to be disabled the wakelock is released.

Change-Id: I72e1fb80e47e9bc181766dc84bafb006a71ff6b4
Signed-off-by: Ulf Hansson <ulf.hansson@stericsson.com>
Reviewed-on: http://gerrit.lud.stericsson.com/gerrit/14668
Reviewed-by: Stefan NILSSON9 <stefan.xk.nilsson@stericsson.com>
Reviewed-by: QATOOLS
Reviewed-by: Jonas ABERG <jonas.aberg@stericsson.com>
Reviewed-on: http://gerrit.lud.stericsson.com/gerrit/30726