drivers/gpu: Include mali through a symlink
The mali device driver is maintained in its own git and has been built
as a kernel module in the past. This commit adds a symlink in
drivers/gpu to the mali400ko git and modifies the Makefile and Kconfig
to include the mali driver in the kernel.

ST-Ericsson ID: CR283174

Change-Id: I7a35ee5c4463ac68f13dc2e45aa77ad18f8e5dc6
Signed-off-by: Magnus Wendt <magnus.wendt@stericsson.com>
Signed-off-by: Jonas Aaberg <jonas.aberg@stericsson.com>
Reviewed-on: http://gerrit.lud.stericsson.com/gerrit/29943
Reviewed-by: Robert FEKETE <robert.fekete@stericsson.com>
Tested-by: Robert FEKETE <robert.fekete@stericsson.com>
Reviewed-by: Ushit KUMAR <ushit.kumar@stericsson.com>
Tested-by: Ushit KUMAR <ushit.kumar@stericsson.com>