MMCI: Increase HW flow control workaround
There is a HW pecularity when writing packets with size <= 8 bytes
to PL18X. Previously HW flow control was disabled for sizes < 8 bytes,
but it needs to be disabled also for sizes == 8 bytes.

Since HW flow control is not really needed for anything that fits in the
HW FIFO of PL18X, we now instead disable HW flow control for any write
operation that is smaller than or equal to the FIFO size.

ST-Ericsson ID: ER326261

Change-Id: Ifa27d47a4c79b2c4542bba5cf2cfc30297403cd0
Signed-off-by: Stefan Nilsson XK <stefan.xk.nilsson@stericsson.com>
Reviewed-on: http://gerrit.lud.stericsson.com/gerrit/17137
Reviewed-by: QATOOLS
Reviewed-by: Ulf HANSSON <ulf.hansson@stericsson.com>
Reviewed-by: Bartosz MARKOWSKI <bartosz.markowski@tieto.com>
Reviewed-by: Sebastian RASMUSSEN <sebastian.rasmussen@stericsson.com>