i2c-nomadik: Add code to retry on timeout failure and set i2c clock to 400KHz
It is seen that i2c-nomadik controller randomly stops
generating the interrupts leading to a i2c timeout.
As a workaround to this problem add retries, to the on going
transfer on failure.
Since this behaviour does not get changed with a wait for
longer duration, the timeout can be reduced to a lower
value, leading to a quicker retry attempt.

ST-Ericsson ID: ER 325661
Signed-off-by: Virupax Sadashivpetimath <virupax.sadashivpetimath@stericsson.com>

Change-Id: Ic9db2b15ff5ce7e656cbbdfeb19270da9e32f9d1
Reviewed-on: http://gerrit.lud.stericsson.com/gerrit/16972
Tested-by: Virupax SADASHIVPETIMATH <virupax.sadashivpetimath@stericsson.com>
Reviewed-by: Jonas ABERG <jonas.aberg@stericsson.com>