smsc911x: use shifted register access
At least one of the boards using this chip needs a shift in
register access. This adds a "shift" field in the platform data
to have the chip working in all cases without special ifdef.
The approach is similaro to the one used in serial_8250 and other
places.  Performance-wise no difference is reported.

Signed-off-by: Alessandro Rubini <rubini@gnudd.com>
Signed-off-by: Robert Marklund <robert.marklund@stericsson.com>