rtc: add AB5500 RTC driver
Add a driver for the RTC block on the AB5500.  The AB5500 RTC has
similarities with both the AB8500 and the AB3100, so this driver
is made generic in order to replace the other two drivers by adding
some variant specific information.

ST-Ericsson Linux next: -
ST-Ericsson ID: WP332221
ST-Ericsson FOSS-OUT ID: Trivial

Change-Id: I35c7ee245ea68e435c800151dbb8f6e10c7abe5d
Signed-off-by: Rabin Vincent <rabin.vincent@stericsson.com>
Reviewed-on: http://gerrit.lud.stericsson.com/gerrit/19331
Reviewed-by: Vijaya Kumar K-1 <vijay.kilari@stericsson.com>
Reviewed-by: Bibek BASU <bibek.basu@stericsson.com>
Reviewed-by: Srinidhi KASAGAR <srinidhi.kasagar@stericsson.com>