power: ab8500-bm: update usb state change interface to otg notifier
Remove usb_state_changed() exported function. Replace it with otg notifier call.
This makes static_di redundant so it is removed. Moreover, there is no need to
protect usb_state.usb_changed with usb_lock. So usb_lock is removed as well.

Usb state is fabricated based on the reported mA. State is not as accurate as
reading it directly from the usb controller. This should not matter as battery-
manager driver should care more about mA than the state anyway.

Signed-off-by: Mian Yousaf Kaukab <mian.yousaf.kaukab@stericsson.com>