video: mcde: fix SDTV through AB8500 flashes
The patch fixes flashes that are seen when running SDTV through AB8500.
The flashes are symptons of writing to the same buffer as MCDE. This is
because MCDE generates a VCMP interrupt for each field, instead of each
frame.

ST-Ericsson ID: 336212
ST-Ericsson Linux next: Not tested, ER 282779
ST-Ericsson FOSS-OUT ID: Trivial

Change-Id: I84b13bf94e11a7742a0369b237594c565d0a64ea
Signed-off-by: Marcel Tunnissen <Marcel.Tuennissen@stericsson.com>
Reviewed-on: http://gerrit.lud.stericsson.com/gerrit/21217
Tested-by: Marcel TUNNISSEN <marcel.tuennissen@stericsson.com>
Reviewed-by: QATOOLS
Reviewed-by: QATEST
Reviewed-by: Jimmy RUBIN <jimmy.rubin@stericsson.com>
Reviewed-by: Per PERSSON <per.xb.persson@stericsson.com>
Reviewed-by: Jonas ABERG <jonas.aberg@stericsson.com>
Reviewed-by: Dan JOHANSSON <dan.johansson@stericsson.com>