WLAN: CW1200: Power management is implemnted.
Power management is on in this commit. Device is configured
in quiescent mode when idle and in dose mode when operating.
BH always checks if device is awake before trying to access it.
Timeout for putting device back to sleep is 1 second after
last device access.

Verification with WSM_A21.05.0288 shows that device is staying
awake even in quiescent mode. Bug in firmware? To be investigated.

Signed-off-by: Dmitry Tarnyagin <dmitry.tarnyagin@stericsson.com>