mach-ux500: Add SDIO VDD handler for GPIO config
It seems as PL18X does not drive all pins in a correct way when the block
is powered but not initialized. Since CW1200 is sensitive to this (it samples
the DAT2 signal level on bootup to choose between SPI and SDIO mode), it must
be assured that the pins are in a correct state at all times.

This is done by adding a VDD handler for SDI1 which reconfigures all the pins
to GPIO pins with pullups when PL18X is not in an initialized state.

ST-Ericsson ID: 327586
ST-Ericsson FOSS-OUT ID: Trivial

Change-Id: Ic3bbe24f13c32203ad1f5fd3fef22ef25996568e
Signed-off-by: Stefan Nilsson XK <stefan.xk.nilsson@stericsson.com>
Reviewed-on: http://gerrit.lud.stericsson.com/gerrit/18619
Reviewed-by: Sebastian RASMUSSEN <sebastian.rasmussen@stericsson.com>
Reviewed-by: Linus WALLEIJ <linus.walleij@stericsson.com>

Conflicts:

	arch/arm/mach-ux500/board-mop500-sdi.c