regulators: Set SysClkReq6 pin in GPIO mode
Set SysClkReq6 pin in GPIO mode as it is not used and this will
save power on VbatA. Also, as there is a problem with Vsmps2
low power mode on HREFP V2 V7x boards, so the Vsmps2 is forced
in high power mode until that issue is solved.

Note that Vsmps2 is already in high power mode before this
patch is applied because the SysClkReq6 pin is not connected
and pulled high which requests Vsmps2 in high power mode.

ST-Ericsson Linux next: -
ST-Ericsson ID: ER 324615
ST-Ericsson FOSS-OUT ID: Trivial

Change-Id: I1fbd203227bad9e781f618a213016ecf1efdd1c2
Signed-off-by: Bengt Jonsson <bengt.g.jonsson@stericsson.com>
Reviewed-on: http://gerrit.lud.stericsson.com/gerrit/20324
Reviewed-by: Yvan FILLION <yvan.fillion@stericsson.com>

Conflicts:

	arch/arm/mach-ux500/board-mop500.c