usb : storage : remove incorrect setting length
It is incorrect to set short_not_ok when preparing
the request to receive CBW.CBW will be a short-packet
so short_not_ok must not be set.

Mass-storage and file-storage gadgets align the length
to maximum-packet-size when preparing the request to
receive CBW. This is unnecessary and prevents the
controller driver from knowing that a short-packet
is expected.

ST-Ericsson ID: 275473
ST-Ericsson Linux next: 275473
ST-Ericsson FOSS-OUT ID: NA

Change-Id: I0d0116f72337c86fcd8861aa8c3a53afc4d84f7e
Signed-off-by: Mian Yousaf <mian.yousaf.kaukab@stericsson.com>
Signed-off-by: rajaram <rajaram.ragupathy@stericsson.com>
Reviewed-on: http://gerrit.lud.stericsson.com/gerrit/20303
Tested-by: Rajaram REGUPATHY <ragupathy.rajaram@stericsson.com>
Reviewed-by: Jonas ABERG <jonas.aberg@stericsson.com>
Reviewed-by: Praveena NADAHALLY <praveen.nadahally@stericsson.com>

Conflicts:

	drivers/usb/gadget/f_mass_storage.c
	drivers/usb/gadget/file_storage.c
	drivers/usb/gadget/storage_common.c