ARM: ux500: cpuidle_dbg: Show ApIdle/ApSleep last prcmu status
Collect the state from which the prcmu woke the ARM from.
This can be useful to see how often we have a pending interrupt
or if we try to go to sleep, but the prcmu only goes to idle.

ST-Ericsson Linux next: ER338824
ST-Ericsson ID: ER338876
ST-Ericsson FOSS-OUT ID: Trivial

Change-Id: I853ca2381b484994ef784d08f7e9d9f6ca088376
Signed-off-by: Jonas Aaberg <jonas.aberg@stericsson.com>
Reviewed-on: http://gerrit.lud.stericsson.com/gerrit/23096
Reviewed-by: QATEST
Reviewed-by: Rickard ANDERSSON <rickard.andersson@stericsson.com>