video: mcde: Change color convert offset fields
This patch changes the order of the color convert offset in such a way
that they make more sense when setting up the color convert for a
specific platform. Now the offsets are in line with the order in the
multiplication matrix.

ST-Ericsson ID: 338207
ST-Ericsson Linux next: Not tested, ER 282779
ST-Ericsson FOSS-OUT ID: Trivial

Change-Id: Ia39ca4ac52e5c21ff32739de207f0445aaf0e25c
Signed-off-by: Marcel Tunnissen <Marcel.Tuennissen@stericsson.com>
Reviewed-on: http://gerrit.lud.stericsson.com/gerrit/22274
Tested-by: Marcel TUNNISSEN <marcel.tuennissen@stericsson.com>
Reviewed-by: QATEST
Reviewed-by: Linus WALLEIJ <linus.walleij@stericsson.com>

Conflicts:

	arch/arm/mach-ux500/board-pdp-mcde.c