nmk-gpio: fix spurious interrupts with SKE
If the pin is switching to altfunc, and there was an interrupt
installed on it which has been lazy disabled, actually mask the
interrupt to prevent spurious interrupts that would occur while
the pin is under control of the peripheral.  Only SKE does this.

ST-Ericsson Linux next: -
ST-Ericsson ID: ER335433
ST-Ericsson FOSS-OUT ID: Trivial

Change-Id: I9aac362d8bcc63d284ae0d483f0703be71328fd9
Reviewed-on: http://gerrit.lud.stericsson.com/gerrit/23929
Reviewed-by: QATEST
Tested-by: Rabin VINCENT <rabin.vincent@stericsson.com>
Reviewed-by: Srinidhi KASAGAR <srinidhi.kasagar@stericsson.com>