SDIO: Fix deadlock when destroying IRQ work queue
This patch fixes a deadlock in sdio_release_irq. If there was work
scheduled that had not yet claimed the mmc host at the point of calling
sdio_release_irq, then it never could claim the host since the caller
of sdio_release_irq already had claimed the host.

Now the host is released to allow the work queue to flush out any
pending work. A check is also added to make sure IRQ:s are not
enabled as a consequence of this cleanup work.

ST-Ericsson Linux next: NA
ST-Ericsson ID: ER339608
ST-Ericsson FOSS-OUT ID: Trivial

Change-Id: I700719e10c779d69561f0cc66b1df5fb41cbd21e
Signed-off-by: Stefan Nilsson XK <stefan.xk.nilsson@stericsson.com>
Reviewed-on: http://gerrit.lud.stericsson.com/gerrit/24675
Reviewed-by: Ulf HANSSON <ulf.hansson@stericsson.com>
Reviewed-by: QATEST