arm: ux500: prcmu_ac_wake_req workaround
This patch adds a check in prcmu_ac_wake_req that the modem is
awake (in terms of the value in the PRCM_MOD_AWAKE_STATUS
register) after the AC_WAKE_ACK has been received from the PRCMU
FW. If the check fails, a retry is made.

This seems to be necessary, since the modem can generate an
AC_WAKE_ACK, and then still go to sleep.

ST Ericsson ID: 340222
ST Ericsson FOSS-OUT ID: trivial

Change-Id: I01acbc72a376a995d5afd36ad88ecf70da911b63
Signed-off-by: Mattias Nilsson <mattias.i.nilsson@stericsson.com>
Reviewed-on: http://gerrit.lud.stericsson.com/gerrit/25132
Reviewed-by: Maxime COQUELIN <maxime.coquelin-nonst@stericsson.com>
Reviewed-by: Yann GAUTIER <yann.gautier@stericsson.com>
Reviewed-by: QATEST
Reviewed-by: Yvan FILLION <yvan.fillion@stericsson.com>
Tested-by: Yvan FILLION <yvan.fillion@stericsson.com>