Ux500 ASoC:DistortionNoise fix for underrun AB5500
Underrun happens due to system load. This causes the
msp and dma to start the channel again and again.
The fix prevents the call to start the channel once it is
started alredy.The fix also increases the start threshold
of the buffer list so that the duration after which the
usecase hits underrun is improved.

ST-Ericsson Linux next: Not tested, ER 344102
ST-Ericsson ID: 344102
ST-Ericsson FOSS-OUT ID: STETL-FOSS-OUT-10006

Change-Id: I992f83efbf8f48ff652e4c164356eb2e75fd3ac0
Signed-off-by: Shyam Krishnan M <shyam.moni@stericsson.com>
Reviewed-on: http://gerrit.lud.stericsson.com/gerrit/25261
Tested-by: Hemanth P SETHURAM <hemanth.sethuram@stericsson.com>
Reviewed-by: QATOOLS
Reviewed-by: QATEST
Reviewed-by: Srinidhi KASAGAR <srinidhi.kasagar@stericsson.com>