PM / Runtime: Update documentation of interactions with system sleep
The documents describing the interactions between runtime PM and
system sleep generally refer to the model in which the system sleep
state is entered through a global firmware or hardware operation.
As a result, some recommendations given in there are not entirely
suitable for systems in which this is not the case.  Update the
documentation to take the existence of those systems into account.

Signed-off-by: Rafael J. Wysocki <rjw@sisk.pl>
Reviewed-by: Kevin Hilman <khilman@ti.com>
Change-Id: I926b72b01331e563d2e4a1e3b48c9e4c2ce29a05
Reviewed-on: http://gerrit.lud.stericsson.com/gerrit/35336
Reviewed-by: QABUILD
Reviewed-by: Linus WALLEIJ <linus.walleij@stericsson.com>
Tested-by: Ulf HANSSON <ulf.hansson@stericsson.com>