cw1200: Workaround against cyclic requeue of multicast frames.
WSM324 firmware has tendency to requeue multicast frames in a loop,
causing performance drop and high power consumption of the driver.
The workaround blocks attempts to requeue multicasts frames.

Workaround has to be modified for further WSM firmware releases where
this issue is suppose to be fixed with respect to backward compatibility.

Change-Id: I27b66d9acc48b8058efd524a72e6f69c816393cc
Signed-off-by: Dmitry Tarnyagin <dmitry.tarnyagin@stericsson.com>
Reviewed-on: http://gerrit.lud.stericsson.com/gerrit/28209
Reviewed-by: Bartosz MARKOWSKI <bartosz.markowski@tieto.com>
Tested-by: Bartosz MARKOWSKI <bartosz.markowski@tieto.com>
Reviewed-on: http://gerrit.lud.stericsson.com/gerrit/33493
Reviewed-by: Philippe LANGLAIS <philippe.langlais@stericsson.com>