power: ab8500-bm: Changed behaviour of the charging safety timer
The safety timer will not be started until the capacity
reported from the FG algorithm is 100%.
Then we know that the amount of charge that's gone into
the battery is enough for the battery to be full.
If it has not reached end-of-charge before the safety timer
has expired then we know that the battery is overcharged
and charging will be stopped to protect the battery.

ST-Ericsson ID: 358130
ST-Ericsson Linux next: NA
ST-Ericsson FOSS-OUT ID: Trivial

Change-Id: If3cfb49c03235dba09dc35d6b885ff22ea592aa9
Signed-off-by: Johan Palsson <johan.palsson@stericsson.com>
Reviewed-on: http://gerrit.lud.stericsson.com/gerrit/29550
Reviewed-by: QABUILD
Reviewed-by: Jonas ABERG <jonas.aberg@stericsson.com>
Reviewed-by: Karl KOMIEROWSKI <karl.komierowski@stericsson.com>
Reviewed-by: Andrew LYNN <andrew.lynn@stericsson.com>
Tested-by: Andrew LYNN <andrew.lynn@stericsson.com>
Reviewed-on: http://gerrit.lud.stericsson.com/gerrit/30881
Reviewed-by: Arun MURTHY <arun.murthy@stericsson.com>
Tested-by: Arun MURTHY <arun.murthy@stericsson.com>