mbox: optimize tx path and fix bug in rx path
During tx, in the work function check for if any messages are pending to be
transmitted, if so, re-transmit. If in rx path work function is queued twice
then the work is executed only once so the other message will be lost. Hence
check if work has been queued and if so read the next message.

ST-Ericsson Linux next: NA
ST-Ericsson ID: 357703
ST-Ericsson FOSS-OUT ID: Trivial

Change-Id: I7f3a18947e36cae30da8ca391d10ed5d3ac0eb03
Signed-off-by: Arun Murthy <arun.murthy@stericsson.com>
Reviewed-on: http://gerrit.lud.stericsson.com/gerrit/31742
Reviewed-by: Bibek BASU <bibek.basu@stericsson.com>
Tested-by: Bibek BASU <bibek.basu@stericsson.com>