U8500 CM: send a DSP Panic if it's not responding
* Generate a DSP Panic and send it to user process when the
DSP does not respond anymore.
The issue is that when a process is using a DSP (ie has some
components deployed on a DSP) without any activity (no interaction
with the DSP), the driver needs to tell it when the DSP is dead.
This Panic message is there to solve this issue.
* Force wake-up of MMDSP before changing hardware power state
  This a work around for prcmu issue.

ST-Ericsson ID: 359048
ST-Ericsson Linux next: NA
ST-Ericsson FOSS-OUT ID: Trivial
Signed-off-by: Pierre Peiffer <pierre.peiffer@stericsson.com>

Change-Id: I920cc6a8e79e7dc9b39bd77f700a9e0e056b6a81
Reviewed-on: http://gerrit.lud.stericsson.com/gerrit/32584
Reviewed-by: Pierre PEIFFER <pierre.peiffer@stericsson.com>
Tested-by: Pierre PEIFFER <pierre.peiffer@stericsson.com>