dma40: Add support to use lcla area from esram
This patch provides an option of having the lcla in ESRAM
instead of allocating it. The bool value (use_esram_lcla) in the
stedma40_platform_data if set to true, then the lcla address
would be taken from platform resources. This patch does not
enable the feature, as the other dependencies related to
keeping the esram34 regulator on, changes in mcde for releasing
8k from esram bank4 and the changes needed in cmdma
have to be merged.

ST-Ericsson ID: 362066
ST-Ericsson FOSS-OUT ID: Trivial
ST-Ericsson Linux next: NA

Change-Id: I590751f732021c5f2f4da9b7d51ae5dc10087be7
Signed-off-by: Narayanan G <narayanan.gopalakrishnan@stericsson.com>
Reviewed-on: http://gerrit.lud.stericsson.com/gerrit/31123
Reviewed-by: Rabin VINCENT <rabin.vincent@stericsson.com>