ux500: u8500: Change DMA configuration for SD-card
This patch is a mixture of two earlier patches:

By Sebastian Rasmussen:
ux500: Use other logical DMA channel for SD-card

Previously PoP eMMC SDI has been in the same event group as
SD-card SDI, which means that they have been sharing the
same physical DMA channels. Moreover one of those channels
was reserved for other purposes, causing transfers for each
interface to compete for the same physical DMA channel. In
addition there appears to be a HW issue that causes data to
be lost in the middle of DMA transfers, which ended up with
DMAC and MMCI not agreeing on how much data was still to be
transferred, thereby resulting in a hang and eventually in
a crash.

This patch moves SD-card DMA transfers to another event
group which means that it no longer competes for the same
physical DMA channel. This results in that no data is lost
in the transfers and no hang is observed.

By Rabin Vincent:
ux500: mop500: force SD/MMC and MSP2 TX onto different channels

SD/MMC with event line 1 can be in channel 0 or 1.  Force both Rx
and Tx onto channel 0.  Both will anyway not used at the same time,
and with this change channel 1 will be available for MSP2 Tx
without sharing.

Change-Id: I72c254b3258615755fb050595a4d82e4f76d2098
Signed-off-by: Ulf Hansson <ulf.hansson@stericsson.com>
Reviewed-on: http://gerrit.lud.stericsson.com/gerrit/31603