mfd: db8500-prcmu: restore pl011 supply name
pl011 supply name was updated so that amba bus driver can control the
regulator. Currently amba bus driver only enables and disables the supply
on module init and exit. This prevents saving power when the device is not
being used. Renaming the supply to give the driver a chance to control the
supply at fine-grain level.

Change-Id: I85fcc50e95695f38f4b07587606fd6a9df056464
Signed-off-by: Mian Yousaf Kaukab <mian.yousaf.kaukab@stericsson.com>
Reviewed-on: http://gerrit.lud.stericsson.com/gerrit/32126
Reviewed-by: Jonas ABERG <jonas.aberg@stericsson.com>
Tested-by: Jonas ABERG <jonas.aberg@stericsson.com>