musb: host: Wait for TXPKTRDY to clear in host
MUSB_TXCSR_TXPKTRDY will not get cleared if the data is not on
the bus. So, need to wait for it to clear in the DMA completion
callback.As per musb datasheet, generally mode 1 needs to be
used for data length greater than or equal to max packet size.

ST-Ericsson ID:AP 363893
ST-Ericsson Linux next: NA
ST-Ericsson FOSS-OUT ID: NA

Change-Id: I934295798a5fbccf60c6f4004595b625a98098ed
Signed-off-by: <thirupathi.chippakurthy@stericsson.com>
Signed-off-by: Praveena <praveen.nadahally@stericsson.com>
Signed-off-by: rajaram <rajaram.ragupathy@stericsson.com>
Reviewed-on: http://gerrit.lud.stericsson.com/gerrit/32012
Reviewed-by: Rabin VINCENT <rabin.vincent@stericsson.com>