regulators: ab8500: Add support of low voltage battery
Low voltage batteries have a wider voltage range with
lower operating voltages. Some consumers in the platform
may not work with the lower voltages and therefore need
an extra regulator to boost the voltage in this case.

This driver adds support for checking the consumers that
need higher voltage (Vaux1, 2 and 3 regulators, 3 V SIM)
and control the external buck/boost regulator
accordingly.

Note that to utilize the low voltage battery support,
the battery voltage thresholds must be changed. This
applies for the low battery voltage threshold of the
battery manager and the OTP setting for the AB8500
BattOk levels.

ST-Ericsson ID: 282517, 363432
ST-Ericsson Linux next: -
ST-Ericsson FOSS-OUT ID: Trivial

Change-Id: Ife07a622ec9748c027dbbd78b01e4ee7e92629ec
Signed-off-by: Bengt Jonsson <bengt.g.jonsson@stericsson.com>
Reviewed-on: http://gerrit.lud.stericsson.com/gerrit/33616
Reviewed-by: QABUILD