dma40: Handle esram regulator in dma driver
If the esram is used for LCLA region, this patch will hold the regulator
during when the DMA driver is active. The regulator will be disabled on
a suspend and will be enabled again on resume.

The dma driver initcall has been moved to subsys_initcall(), to ensure
that the regulators are available before we do a regulator_get().

ST-Ericsson ID: 364165
ST-Ericsson FOSS-OUT ID: Trivial
ST-Ericsson Linux next: NA

Change-Id: Ieee3c329c26e9fb43f282716fa57810211f79d85
Signed-off-by: Narayanan G <narayanan.gopalakrishnan@stericsson.com>
Reviewed-on: http://gerrit.lud.stericsson.com/gerrit/33434
Reviewed-by: Rabin VINCENT <rabin.vincent@stericsson.com>