power: ab8500_bm: Removal of maintenance charging
This patch makes it possible to remove the maintenance A and B
charging states and replaces them with a state that restarts the
charging when the battery voltage level dropped below a certain
level. The maintenance charging can now be turned of by setting
a flag in the board configuration.

ST-Ericsson Linux Next: -
ST-Ericsson ID: CR333019
ST-Ericsson FOSS-OUT ID: Trivial

Change-Id: Ie5e8ab20a3b57a1028d544371b10ffbb1fd77660
Signed-off-by: Kalle Komierowski <kalle.komierowski@gmail.com>
Reviewed-on: http://gerrit.lud.stericsson.com/gerrit/19916
Reviewed-by: Johan GARDSMARK <johan.gardsmark@stericsson.com>
Reviewed-by: Johan PALSSON <johan.palsson@stericsson.com>
Reviewed-by: Jonas ABERG <jonas.aberg@stericsson.com>
Tested-by: Karl KOMIEROWSKI <karl.komierowski@stericsson.com>

Conflicts:

	arch/arm/mach-ux500/board-mop500-bm.c