mach-ux500: pm: runtime: upgrade to 3.0
Support for changes in:

commit 4d27e9dcff00a6425d779b065ec8892e4f391661
Author: Rafael J. Wysocki <rjw@sisk.pl>
Date:   Fri Apr 29 00:35:50 2011 +0200

    PM: Make power domain callbacks take precedence over subsystem ones

    Change the PM core's behavior related to power domains in such a way
    that, if a power domain is defined for a given device, its callbacks
    will be executed instead of and not in addition to the device
    subsystem's PM callbacks.

The idle->suspend happened earlier via the bus level runtime_idle which
called pm_generic_runtime_idle().  We now need a power domain callback
to do the same thing because of the above commit.

The separate domains for amba and platform are required because the
legacy suspend/resume callbacks need to be called via these specific
functions.

Signed-off-by: Rabin Vincent <rabin.vincent@stericsson.com>