Ugly-rename amba-pl022 to spi-pl022
This ugly renaming of the file amba-pl022 to spi-pl022 is
done in order to be able to apply PL022 patches from the
mainline kernel without hazzle.

Grant reorganized all the drivers but to apply his original
renaming patch we would have to bring in all patches to all
*other* drivers up to that point, which would in turn bring
in a large portion of the kernel 3.1 baseline - let's not
go there.

This is ugly but the alternatives are uglier. When
rebasing to a kernel later than this one, remove this patch
and use mainline commits as far as possible.

Signed-off-by: Linus Walleij <linus.walleij@stericsson.com>
Change-Id: If5a453371f46b09af88febdd03319d1f2fcb612b
Reviewed-on: http://gerrit.lud.stericsson.com/gerrit/34206