power:ab5500-charger: check the usb type in notifier call
When a usb notifier is called, update the notified current only if the plugge
in device is a standard host else do not update the current just return.
The notifier call that is called in the usb driver is not only for the
battery driver but other usb instances, hence usb cannot stop calling the
notifier call and do a check to call only if it is a standard host. Hence
this check is done in the battery driver, and if relevant update else just
return.

ST-Ericsson Linux next: NA
ST-Ericsson ID: 362955
ST-Ericsson FOSS-OUT ID: Trivial

Change-Id: I83026aba60b8a60c9f80e2a699b8eded285f8a09
Signed-off-by: Arun Murthy <arun.murthy@stericsson.com>
Reviewed-on: http://gerrit.lud.stericsson.com/gerrit/34799
Reviewed-by: Mian Yousaf KAUKAB <mian.yousaf.kaukab@stericsson.com>
Reviewed-by: QABUILD
Reviewed-by: Srinidhi KASAGAR <srinidhi.kasagar@stericsson.com>
Reviewed-by: Praveena NADAHALLY <praveen.nadahally@stericsson.com>