AMBA: Use suspend_noriq to force devices into runtime suspend
To be able to make sure devices are put into runtime suspend
after a suspend sequence, the suspend_noirq callback is used.

Previously it was was possible for drivers doing pm_runtime_suspend
and pm_runtime_put_sync directly from it's suspend callbacks.
This is now not possible due to the following commit, which solve a
race issue:

PM: Limit race conditions between runtime PM and system sleep (v2)

Change-Id: I48b988e9f7e411a0a122a58c0967fbee61256f89
Signed-off-by: Ulf Hansson <ulf.hansson@stericsson.com>
Reviewed-on: http://gerrit.lud.stericsson.com/gerrit/35505
Reviewed-by: Linus WALLEIJ <linus.walleij@stericsson.com>