mmci: add constraints on alignment for SDIO
Buffers must be 4 bytes aligned due to restrictions that
the PL18x FIFO accesses must be done in a 4-byte aligned manner.
Enable DMA_REQCTL for SDIO to support write of not 32 bytes aligned
sg element lengths. In PIO mode any buffer length can be handled
as long as the buffer address is 4 byte aligned.

Signed-off-by: Ulf Hansson <(address hidden)>
Signed-off-by: Per Forlin <(address hidden)>
Signed-off-by: Stefan Nilsson XK <(address hidden)>