cw1200: Revert requesting SDIO pins at device registration
The SDIO pins belongs to the SDIO driver and
should be handled by the SDIO driver only.

The reason that it has worked up until now is that
the SDIO driver has just started using pm runtime.
Since both WLAN and SDIO framework enable
the pins, the number of users is set to two.

SDIO tries to release the pins (after 50 ms of inactivity
in order to save power), the usage count does not go
down to 0 and therefore the pins are not reconfigured
to GPIOs, which they must be when PL18X is not driving
the pins due to power down).

This reverts commit 53c8821cafaabc841d194c7d86cca3b574605e6e

Signed-off-by: Bartosz Markowski <bartosz.markowski@tieto.com>
Signed-off-by: Philippe Langlais <philippe.langlais@stericsson.com>