rtc: disable the alarm in the hardware
Currently, the RTC code does not disable the alarm in the hardware.

This means that after a sequence such as the one below (the files are in the
RTC sysfs), the box will boot up after 2 minutes even though we've
asked for the alarm to be turned off.

	# echo $((`cat since_epoch`)+120) > wakealarm
	# echo 0 > wakealarm
	# poweroff

Fix this by disabling the alarm when there are no timers to run.

Change-Id: Ie9bddb60afc924265d1fb1562a4d6df61faf1d61
Signed-off-by: Rabin Vincent <rabin.vincent@stericsson.com>

ST-Ericsson ID: 375473
ST-Ericsson Linux next: NA
ST-Ericsson FOSS-OUT ID: Trivial

Change-Id: Ia0246639c2bc30ce99264b4d5ccdc4c003a13b18
Reviewed-on: http://gerrit.lud.stericsson.com/gerrit/39115
Reviewed-by: QABUILD
Tested-by: Rabin VINCENT <rabin.vincent@stericsson.com>
Reviewed-by: Jonas ABERG <jonas.aberg@stericsson.com>