musb: host: Wait for TXPKTRDY to clear in host
MUSB_TXCSR_TXPKTRDY will not get cleared if the data is not on
the bus. So, need to wait for it to clear in the DMA completion
callback.As per musb datasheet, generally mode 1 needs to be
used for data length greater than or equal to max packet size.
Bus reset happening due to scsi timeout,since in this case
should not return continue for next transfers.

ST-Ericsson ID:ER 365087
ST-Ericsson Linux next: NA
ST-Ericsson FOSS-OUT ID: NA

Signed-off-by: Thirupathi <thirupathi.chippakurthy@stericsson.com>