regulator: db5500: Correct propagation to power state
This patch solves an issue with the power state count
(which is used when CPUidle takes the decision to go to
sleep).

When a db5500 regulator is enabled/disabled, the power
state count should only be updated if there is a change
in the regulator state. Normally, this error is not
visible because the regulator framework will check if
the regulator is enabled before enabling. But if the
suspend/resume functionality of the regulator framework
is activated, there may be calls to enable even if the
regulator is already enabled.

ST-Ericsson ID: 398507
ST-Ericsson Linux next: NA
ST-Ericsson FOSS-OUT ID: Trivial

Change-Id: Ic9a875c6c00119f470ca76de2cedce8204b9cbb8
Signed-off-by: Bengt Jonsson <bengt.g.jonsson@stericsson.com>
Reviewed-on: http://gerrit.lud.stericsson.com/gerrit/39924
Reviewed-by: QABUILD
Reviewed-by: Vijaya Kumar K-1 <vijay.kilari@stericsson.com>
Reviewed-by: Rabin VINCENT <rabin.vincent@stericsson.com>
Tested-by: Rabin VINCENT <rabin.vincent@stericsson.com>