regulator: db8500: Correct propagation to power state
This patch solves an issue with the power state count
(which is used when CPUidle takes the decision to go to
sleep).

When a db8500 regulator is enabled/disabled, the power
state count should only be updated if there is a change
in the regulator state. Normally, this error is not
visible because the regulator framework will check if
the regulator is enabled before enabling. But if the
suspend/resume functionality of the regulator framework
is activated, there may be calls to enable even if the
regulator is already enabled.

ST-Ericsson ID: 398507
ST-Ericsson Linux next: NA
ST-Ericsson FOSS-OUT ID: Trivial

Change-Id: Ieee237877457d6f8ea37b2d9a873d4dcaebf7e3d
Signed-off-by: Bengt Jonsson <bengt.g.jonsson@stericsson.com>
Reviewed-on: http://gerrit.lud.stericsson.com/gerrit/39921
Reviewed-by: QATEST
Reviewed-by: QABUILD
Reviewed-by: Jonas ABERG <jonas.aberg@stericsson.com>