u8500-shrm: add timer to check if modem has stuck or hang
On APE sending msg_pend notification, the only way when APE gets to know that
modem is not responsive, is when it ends up with a FIFO full. In order to get
this FIFO full on common channel it might take 30min. Hence add a timer and
start the timer on msg_pend_notification and cancel the timer at read_noti.
If timer expires within this time then its for sure that modem is not
responsive. Initiate MSR in that case.
The current timer value is 6sec.

Since there are two timers and on expiry of either of the timer, modem_reset
is called, have added flag so as to ensure that modem_rest is not called
simultaneously on exit of both the timer in short span of time.

ST-Ericsson Linux next: NA
ST-Ericsson ID: 407459
ST-Ericsson FOSS-OUT ID: Trivial

Change-Id: If3e7b61ee57ea4b1d386a50c8c4b99c2cd701e13
Signed-off-by: Arun Murthy <arun.murthy@stericsson.com>
Reviewed-on: http://gerrit.lud.stericsson.com/gerrit/46337
Reviewed-by: Ashok G X <ashok.g@stericsson.com>
Tested-by: Ashok G X <ashok.g@stericsson.com>
Reviewed-by: Srinidhi KASAGAR <srinidhi.kasagar@stericsson.com>