gpio/stmpe: Change implementation of noirq mode
Removes the use of the new no_irq platform data member and instead uses
the value obtained for the cell irq to know if irqs are supported or
not. If irqs are not supported, the device will not have been given an
irq resource and the irq value will be invalid.

ST-Ericsson ID: 399362
ST-Ericsson FOSS-OUT ID: Trivial
ST-Ericsson Linux next: NA

Change-Id: Ib7e88229b58178952405074295c70471ea1a0681
Signed-off-by: Chris Blair <chris.blair@stericsson.com>
Reviewed-on: http://gerrit.lud.stericsson.com/gerrit/46690
Reviewed-by: QATOOLS
Reviewed-by: Linus WALLEIJ <linus.walleij@stericsson.com>