Ux500 CM: avoid ugly warning if kmalloc() fails
In some cases, if the CM driver fails to allocate memory through
kmalloc(), it falls back into vmalloc() which has better chance to
succeed for big allocations. When this happens, kmalloc() reports its
failure with an ugly warning on the console. Just avoid this warning
in this case.

ST-Ericsson ID: 413249
ST-Ericsson FOSS-OUT ID: Trivial
ST-Ericsson Linux next: N/A

Change-Id: I6c8a51eb430891727854b1800cf3a649dcb4b6e9
Signed-off-by: Pierre Peiffer <pierre.peiffer@stericsson.com>
Reviewed-on: http://gerrit.lud.stericsson.com/gerrit/46815
Reviewed-by: QATOOLS