cpufreq:ondemand: Handle SW coordinated CPUs
This patch fixes a bug that occured when we had load on
a secondary CPU and the primary CPU was sleeping. Only
one sampling timer was spawned and it was spawned as a
deferred timer on the primary CPU, so when a secondary CPU
had a change in load this was not detected by the ondemand
governor.
This patch make sure that deferred timers are run on all
CPUs in the case of software controlled CPUs that run on
the same frequency.

ST-Ericsson ID: 410219
ST-Ericsson Linux next: NA
ST-Ericsson FOSS-OUT ID: Trivial

Change-Id: I9875dd5b477e5b8d1b51438d5355ad4c7c58dba1
Signed-off-by: Rickard Andersson <rickard.andersson@stericsson.com>
Reviewed-on: http://gerrit.lud.stericsson.com/gerrit/48218
Reviewed-by: QABUILD
Reviewed-by: Mian Yousaf KAUKAB <mian.yousaf.kaukab@stericsson.com>
Reviewed-by: Jonas ABERG <jonas.aberg@stericsson.com>