u8500-shrm: secure the access of GOP register
Before accessing the GOP registers checks are present to verify the modem
reset status and the availability of hostaccess_port. There are chances
of preemtion happening between the check and access to GOP during which
if modem has reset, then we will end up with a system freeze.
Hence secure that the check and access to GOP happends at a time without
preemption.

ST-Ericsson Linux next: NA
ST-Ericsson ID: 413508
ST-Ericsson FOSS-OUT ID: Trivial

Signed-off-by: Arun Murthy <arun.murthy@stericsson.com>
Change-Id: I481e925b3a457ad3c8c82a5b7986dba6b2a4457b
Reviewed-on: http://gerrit.lud.stericsson.com/gerrit/48276
Reviewed-by: QABUILD
Reviewed-by: Srinidhi KASAGAR <srinidhi.kasagar@stericsson.com>