musb: Reselect index reg in interrupt context
musb INDEX register is getting corrupted due to locking issues.
So, set this register with proper value after acquiring the lock

Simple example of this scenario:
Consider an interrupt and a thread running on SMP system. Each one of
these are running on different CPU's.  A global variable x is shared
across these two
CPU1: Interrupt starts execution
CPU1: Acquires a LOCK
CPU1: Interrupt sets x=1
CPU1: ..
CPU1: ..
CPU1: Interrupt releases the LOCK (for specific reason) and continues
execution
CPU2: Thread acquires the LOCK
CPU2: Thread sets x=2
CPU1:..
CPU2:..
CPU1:.
CPU2:.
CPU1: Interrupt tries to acquire the LOCK. But, LOCK is not available as
Thread
running on CPU2 has already acquired it. Interrupt spins.
CPU2: .
CPU2: Finishes the work and releases the LOCK
CPU1: Interrupt acquires the LOCK and starts using x assuming that x
value is still 1 which Is wrong

Details of this issue with respect to musb code:
CPU1: Data transfer completion interrupt is received for the class1 EP
CPU1: MUSB INDEX register is set to the Class1 EP number
CPU1: MUSB LOCK is acquired
CPU2: Class2 thread queues the data for transfer.
CPU2: Class2 thread tries to acquire the MUSB LOCK. But fails to
acquire since the lock is already taken by the Class1 running on
CPU1. So, Class2 thread spins for the lock
CPU1: Some work related to Class1 EP is done giveback
function is called.
CPU1: Inside the giveback function, the MUSB LOCK is released
CPU2: Class2 which was spinning for the lock, acquires it now.
CPU1: Class1 Request's completion callback is called
CPU1: Interrupt for Class1 EP tries to acquire the MUSB LOCK. But
fails to acquire since the lock is already taken by the Class2
running on CPU2. So, Interrupt for Class1 EP spins for the lock
CPU2: Class2 schedules the data transfer and sets the MUSB INDEX to
Class2 EP number
CPU2: Class2 completes the scheduling etc and releases the MUSB LOCK
CPU1: Interrupt for Class1 EP acquires the MUSB LOCK.
CPU1: Interrupt for Class1 EP schedules the next data transfer
assuming that MUSB INDEX is not changed. But this is not true as the
MSUB INDEX register was already changed by the Class2 thread

Since the MUSB INDEX register is set to a different endpoint we
read and modify the wrong registers. Hence data transfer will not
happen properly. This results in unpredictable behavior

So, the MUSB INDEX register is set to proper value again when
interrupt re-acquires the lock

ST-Ericsson Linux next: NA
ST-Ericsson ID: 403908
ST-Ericsson FOSS-OUT ID: Trivial

Change-Id: Ie73b37b8104ba7d36040f3bd9fe101b9d898e00d
Signed-off-by: supriya karanth <supriya.karanth@stericsson.com>
Reviewed-on: http://gerrit.lud.stericsson.com/gerrit/48448
Reviewed-by: QATOOLS
Reviewed-by: QABUILD
Reviewed-by: QATEST
Reviewed-by: Praveena NADAHALLY <praveen.nadahally@stericsson.com>
Reviewed-by: Srinidhi KASAGAR <srinidhi.kasagar@stericsson.com>