usb/host: Add ISP1763 host controller support
This is a driver for the ISP1763 USB host controller from ST-Ericsson,
modified by Calao and imported to the kernel tree by Movial.

Only hardware tested is the AVI daughterboard for Snowball from Calao,
although the driver includes support for PCI evaluation version as well.

Note that the driver depends on changes to the kernel USB interfaces
from another patch, and needs a big hammer to iron out the warts in it.