cw1200: Add workaround for WFD test case 6.1.10
Adds workaround for WFD test case 6.1.10.
The test engine misbehaves and sends ACTION
(Device Discoverability) frames with the same sequence number which
is then dropped by the firmware.
To get around this we temporarily MAP to SA of the ACTION frame
and sent a reset request to the FW.

ST-Ericsson ID: 372706
ST-Ericsson FOSS-OUT ID: NA
Signed-off-by: Ajitpal.Singh <ajitpal.singh@stericsson.com>

Change-Id: Id1d80920a3d3c57d14a81e771ff5b0dc73a4184d
Reviewed-on: http://gerrit.lud.stericsson.com/gerrit/37167
Reviewed-by: Bartosz MARKOWSKI <bartosz.markowski@tieto.com>
Tested-by: Ajit Pal SINGH <ajitpal.singh@stericsson.com>