mfd: ab8500: don't use mutex_lock_interruptible
This disadvantage of having it interrptible is that it would
appear to work OK in most situations, but in the rare case
that this call does fail, it can lead to misconfiguration
of the AB, such as a regulator not being turned off and
leading to a platform power consumption increase.

There's no real benefit to having this interruptible.

Make it a plain mutex_lock.  The non-interruptability matches
other low-level I/O functions such as SPI and I2C.

ST-Ericsson ID: 420727
ST-Ericsson Linux next: NA
ST-Ericsson FOSS-OUT ID: Trivial

Change-Id: I38f325f1d6e24afc5f8def0f070aad2b79cb1a34
Signed-off-by: Rabin Vincent <rabin.vincent@stericsson.com>
Reviewed-on: http://gerrit.lud.stericsson.com/gerrit/51048