ARM: smp_twd: Only wait for reprogramming on active cpus
During booting of cpu1, there is a short window where cpu1
is online, but not active where cpu1 is occupied by waiting
to become active. If cpu0 then decides to schedule something
on cpu1 and wait for it to complete, before cpu0 has set
cpu1 active, we have a deadlock.

This is just a work around to hide the original problem.

ST-Ericsson Linux next: -
ST-Ericsson ID: 419957
ST-Ericsson FOSS-OUT ID: Trivial

Change-Id: I133e511cece4024375cc92a82f79bd799514d71d
Signed-off-by: Jonas Aaberg <jonas.aberg@stericsson.com>
Reviewed-on: http://gerrit.lud.stericsson.com/gerrit/51514
Reviewed-by: QABUILD
Reviewed-by: Rickard ANDERSSON <rickard.andersson@stericsson.com>